+++ lib/ansible/modules/cloud/digital_ocean/digital_ocean_tag.py
@@ -103,85 +103,19 @@ data:
     }
 '''
 
-import json
-import os
-
+from traceback import format_exc
 from ansible.module_utils.basic import AnsibleModule
-from ansible.module_utils.urls import fetch_url
-from ansible.module_utils._text import to_text
-
-
-class Response(object):
-
-    def __init__(self, resp, info):
-        self.body = None
-        if resp:
-            self.body = resp.read()
-        self.info = info
-
-    @property
-    def json(self):
-        if not self.body:
-            if "body" in self.info:
-                return json.loads(to_text(self.info["body"], errors='surrogate_or_strict'))
-            return None
-        try:
-            return json.loads(to_text(self.body, errors='surrogate_or_strict'))
-        except ValueError:
-            return None
-
-    @property
-    def status_code(self):
-        return self.info["status"]
-
-
-class Rest(object):
-
-    def __init__(self, module, headers):
-        self.module = module
-        self.headers = headers
-        self.baseurl = 'https://api.digitalocean.com/v2'
-
-    def _url_builder(self, path):
-        if path[0] == '/':
-            path = path[1:]
-        return '%s/%s' % (self.baseurl, path)
-
-    def send(self, method, path, data=None):
-        url = self._url_builder(path)
-        data = self.module.jsonify(data)
-
-        resp, info = fetch_url(self.module, url, data=data, headers=self.headers, method=method)
-
-        return Response(resp, info)
-
-    def get(self, path, data=None):
-        return self.send('GET', path, data)
-
-    def put(self, path, data=None):
-        return self.send('PUT', path, data)
-
-    def post(self, path, data=None):
-        return self.send('POST', path, data)
-
-    def delete(self, path, data=None):
-        return self.send('DELETE', path, data)
+from ansible.module_utils.digital_ocean import DigitalOceanHelper
+from ansible.module_utils._text import to_native
 
 
 def core(module):
-    try:
-        api_token = module.params['api_token'] or \
-            os.environ['DO_API_TOKEN'] or os.environ['DO_API_KEY']
-    except KeyError as e:
-        module.fail_json(msg='Unable to load %s' % e.message)
-
     state = module.params['state']
     name = module.params['name']
     resource_id = module.params['resource_id']
     resource_type = module.params['resource_type']
 
-    rest = Rest(module, {'Authorization': 'Bearer {0}'.format(api_token),
-                         'Content-type': 'application/json'})
+    rest = DigitalOceanHelper(module)
 
     # Check if api_token is valid or not
     response = rest.get('account')
@@ -274,7 +208,7 @@ def main():
     try:
         core(module)
     except Exception as e:
-        module.fail_json(msg=str(e))
+        module.fail_json(msg=to_native(e), exception=format_exc())
 
 
 if __name__ == '__main__':
